pyproject.toml
[project]
name = "neocortex"
version = "0.1.0"
description = "Neocortex multi-market prompt-driven equity research platform"
readme = "DESIGN.md"
requires-python = ">=3.11"
dependencies = [
    "akshare>=1.18.40",
    "baostock>=0.8.9",
    "efinance>=0.5.8",
    "httpx>=0.28.1",
    "jinja2>=3.1.6",
    "lark-oapi>=1.5.3",
    "pyyaml>=6.0.3",
    "SQLAlchemy>=2.0.41",
]

[project.scripts]
neocortex = "neocortex.cli:main"

[dependency-groups]
dev = [
    "pytest>=9.0.0",
    "ruff>=0.12.0",
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.package-data]
"neocortex.prompts" = ["templates/*.yaml"]

[tool.setuptools.packages.find]
where = ["src"]
